Just saving a PQ demo, before refactoring.

+++ src/distance.rs
@@ -1,6 +1,6 @@
 use ndarray::{ArrayView1};
 
-pub fn dist_euclidian(p: &ArrayView1::<f32>, q: &ArrayView1::<f32>) -> f32 {
+pub fn dist_euclidian(p: &ArrayView1::<f64>, q: &ArrayView1::<f64>) -> f64 {
     let mut sum_val = 0.0;
     for i in 0..p.len() {
         sum_val += (p[i]-q[i]).powi(2);
@@ -8,7 +8,7 @@ pub fn dist_euclidian(p: &ArrayView1::<f32>, q: &ArrayView1::<f32>) -> f32 {
     return sum_val.sqrt();
 }
 
-pub fn dist_cosine_similarity(p: &ArrayView1::<f32>, q: &ArrayView1::<f32>) -> f32 {
+pub fn dist_cosine_similarity(p: &ArrayView1::<f64>, q: &ArrayView1::<f64>) -> f64 {
     let dot_prod = p.dot(q);
     let magnitude_p = p.dot(p).sqrt();
     let magnitude_q = q.dot(q).sqrt();
@@ -16,12 +16,12 @@ pub fn dist_cosine_similarity(p: &ArrayView1::<f32>, q: &ArrayView1::<f32>) -> f
     return cos_sim;
 }
 
-pub fn dist_angular_similarity(p: &ArrayView1::<f32>, q: &ArrayView1::<f32>) -> f32 {
+pub fn dist_angular_similarity(p: &ArrayView1::<f64>, q: &ArrayView1::<f64>) -> f64 {
     let dot_prod = p.dot(q);
     let magnitude_p = p.dot(p).sqrt();
     let magnitude_q = q.dot(q).sqrt();
     let cos_sim = dot_prod / (magnitude_p*magnitude_q);
-    return cos_sim.acos() / std::f32::consts::PI;
+    return cos_sim.acos() / std::f64::consts::PI;
 }
 
 pub enum DistType {
